graphql-gateway/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql
httpx